debt - introduce editor.ts in browser namespace
