+++ src/vs/workbench/browser/parts/editor/baseEditor.ts
@@ -5,16 +5,12 @@
 'use strict';
 
 import { TPromise } from 'vs/base/common/winjs.base';
-import types = require('vs/base/common/types');
 import { Builder } from 'vs/base/browser/builder';
-import { Registry } from 'vs/platform/registry/common/platform';
 import { Panel } from 'vs/workbench/browser/panel';
-import { EditorInput, EditorOptions, IEditorDescriptor, IEditorRegistry } from 'vs/workbench/common/editor';
+import { EditorInput, EditorOptions } from 'vs/workbench/common/editor';
 import { IEditor, Position } from 'vs/platform/editor/common/editor';
-import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
-import { IConstructorSignature0 } from 'vs/platform/instantiation/common/instantiation';
 
 /**
  * The base class of editors in the workbench. Editors register themselves for specific editor inputs.
@@ -124,146 +120,3 @@ export abstract class BaseEditor extends Panel implements IEditor {
 		super.dispose();
 	}
 }
-
-/**
- * A lightweight descriptor of an editor. The descriptor is deferred so that heavy editors
- * can load lazily in the workbench.
- */
-export class EditorDescriptor extends SyncDescriptor<BaseEditor> implements IEditorDescriptor {
-	private id: string;
-	private name: string;
-
-	constructor(ctor: IConstructorSignature0<BaseEditor>, id: string, name: string) {
-		super(ctor);
-
-		this.id = id;
-		this.name = name;
-	}
-
-	public getId(): string {
-		return this.id;
-	}
-
-	public getName(): string {
-		return this.name;
-	}
-
-	public describes(obj: any): boolean {
-		return obj instanceof BaseEditor && (<BaseEditor>obj).getId() === this.id;
-	}
-}
-
-const INPUT_DESCRIPTORS_PROPERTY = '__$inputDescriptors';
-
-class EditorRegistry implements IEditorRegistry {
-	private editors: EditorDescriptor[];
-
-	constructor() {
-		this.editors = [];
-	}
-
-	public registerEditor(descriptor: EditorDescriptor, editorInputDescriptor: SyncDescriptor<EditorInput>): void;
-	public registerEditor(descriptor: EditorDescriptor, editorInputDescriptor: SyncDescriptor<EditorInput>[]): void;
-	public registerEditor(descriptor: EditorDescriptor, editorInputDescriptor: any): void {
-
-		// Support both non-array and array parameter
-		let inputDescriptors: SyncDescriptor<EditorInput>[] = [];
-		if (!types.isArray(editorInputDescriptor)) {
-			inputDescriptors.push(editorInputDescriptor);
-		} else {
-			inputDescriptors = editorInputDescriptor;
-		}
-
-		// Register (Support multiple Editors per Input)
-		descriptor[INPUT_DESCRIPTORS_PROPERTY] = inputDescriptors;
-		this.editors.push(descriptor);
-	}
-
-	public getEditor(input: EditorInput): EditorDescriptor {
-		const findEditorDescriptors = (input: EditorInput, byInstanceOf?: boolean): EditorDescriptor[] => {
-			const matchingDescriptors: EditorDescriptor[] = [];
-
-			for (let i = 0; i < this.editors.length; i++) {
-				const editor = this.editors[i];
-				const inputDescriptors = <SyncDescriptor<EditorInput>[]>editor[INPUT_DESCRIPTORS_PROPERTY];
-				for (let j = 0; j < inputDescriptors.length; j++) {
-					const inputClass = inputDescriptors[j].ctor;
-
-					// Direct check on constructor type (ignores prototype chain)
-					if (!byInstanceOf && input.constructor === inputClass) {
-						matchingDescriptors.push(editor);
-						break;
-					}
-
-					// Normal instanceof check
-					else if (byInstanceOf && input instanceof inputClass) {
-						matchingDescriptors.push(editor);
-						break;
-					}
-				}
-			}
-
-			// If no descriptors found, continue search using instanceof and prototype chain
-			if (!byInstanceOf && matchingDescriptors.length === 0) {
-				return findEditorDescriptors(input, true);
-			}
-
-			if (byInstanceOf) {
-				return matchingDescriptors;
-			}
-
-			return matchingDescriptors;
-		};
-
-		const descriptors = findEditorDescriptors(input);
-		if (descriptors && descriptors.length > 0) {
-
-			// Ask the input for its preferred Editor
-			const preferredEditorId = input.getPreferredEditorId(descriptors.map(d => d.getId()));
-			if (preferredEditorId) {
-				return this.getEditorById(preferredEditorId);
-			}
-
-			// Otherwise, first come first serve
-			return descriptors[0];
-		}
-
-		return null;
-	}
-
-	public getEditorById(editorId: string): EditorDescriptor {
-		for (let i = 0; i < this.editors.length; i++) {
-			const editor = this.editors[i];
-			if (editor.getId() === editorId) {
-				return editor;
-			}
-		}
-
-		return null;
-	}
-
-	public getEditors(): EditorDescriptor[] {
-		return this.editors.slice(0);
-	}
-
-	public setEditors(editorsToSet: EditorDescriptor[]): void {
-		this.editors = editorsToSet;
-	}
-
-	public getEditorInputs(): any[] {
-		const inputClasses: any[] = [];
-		for (let i = 0; i < this.editors.length; i++) {
-			const editor = this.editors[i];
-			const editorInputDescriptors = <SyncDescriptor<EditorInput>[]>editor[INPUT_DESCRIPTORS_PROPERTY];
-			inputClasses.push(...editorInputDescriptors.map(descriptor => descriptor.ctor));
-		}
-
-		return inputClasses;
-	}
-}
-
-export const Extensions = {
-	Editors: 'workbench.contributions.editors'
-};
-
-Registry.add(Extensions.Editors, new EditorRegistry());
